+++ control/src/pages/settings/components/SettingsTab.js
@@ -4,7 +4,7 @@ import styles from '../styles/SettingsTab.module.css'
 import TimeZone from './TimeZone'
 import { Link } from 'react-router-dom'
 import PreferenceWrapper from './PreferenceWrapper'
-import { AnimateSharedLayout } from 'framer-motion'
+// import { AnimateSharedLayout } from 'framer-motion'
 import SavePassword from './SavePassword'
 import TwoFactor from './TwoFactor'
 import ChangeEmail from './ChangeEmail'
