+++ control/src/pages/signout/index.js
@@ -1,13 +1,45 @@
 import React, { useEffect, useState } from 'react'
 import { withRouter } from 'react-router-dom'
 import axios from 'axios'
-// import {GetUserInfo} from '../../zuri-control'
 import { SignoutStyleWrapper } from '../../component-styles/SignoutStyle'
 import logo from '../../component-assets/zuri.svg'
 import { Helmet } from 'react-helmet'
 
 const Signout = ({ history }) => {
   const orgName = localStorage.getItem('orgName')
+
+
+  const currentWorkspace = localStorage.getItem('currentWorkspace')
+  let token = sessionStorage.getItem('token')
+
+  const logout = () => {
+    axios({
+      method: 'post',
+      url: `https://api.zuri.chat/auth/logout`,
+      headers: {
+        Authorization: `Bearer ${token}`
+      }
+    })
+      .then(res => {
+        console.log(res)
+      })
+      .catch(err => {
+        console.error(err)
+      })
+  }
+  useEffect(() => {
+    let user = JSON.parse(sessionStorage.getItem('user'))
+
+    if ((user && token) !== null) {
+       logout();
+    
+    } else {
+      console.log('YOU ARE NOT LOGGED IN, PLEASE LOG IN')
+    }
+  }, [orgName])
+
+ 
+ 
 
   return (
     <>
