+++ package.json
@@ -83,6 +83,7 @@
     "react-router-dom": "^5.3.0",
     "react-timezone-select": "^1.0.8",
     "rxjs": "^7.3.0",
-    "single-spa": "^5.9.3"
+    "single-spa": "^5.9.3",
+    "style-value-types": "^5.0.0"
   }
 }
