+++ topbar/src/components/TopbarModal.js
@@ -14,7 +14,6 @@ import MembersModal from './MembersModal'
 import Downloads from './Downloads'
 import PauseNotification from './PauseNotification'
 import SetStatusModal from './SetStatusModal'
-import ProfilePicView from './ProfilePicView'
 // react icons
 
 const TopbarModal = ({ members }) => {
@@ -22,11 +21,10 @@ const TopbarModal = ({ members }) => {
     useContext(ProfileContext)
 
   const state = useContext(TopbarContext)
-  const [showModal, setShowModal] = state.show
+  const [showModal] = state.show
   // const [username, setUsername] = state.username
   const [showStatus] = state.status
   const [showMembersModal] = state.modal
-
   const {
     onEmojiClick,
     openModal,
@@ -38,58 +36,42 @@ const TopbarModal = ({ members }) => {
     reusableModal,
     setReusableModal,
     presence,
-    setPresence,
-    profilePicView,
-    setProfilePicView
+    setPresence
   } = state
+
 
   const currentWorkspace = localStorage.getItem('currentWorkspace')
   let token = sessionStorage.getItem('token')
+  
 
-  useEffect(() => {
-    let user = JSON.parse(sessionStorage.getItem('user'))
+  // useEffect(() => {
+  //   let user = JSON.parse(sessionStorage.getItem('user'))
 
-    if ((user && token) !== null) {
-      try {
-        axios
-          .get(`https://api.zuri.chat/organizations/${currentWorkspace}`, {
-            headers: {
-              Authorization: `Bearer ${token}`
-            }
-          })
-          .then(response => {
-            localStorage.setItem('orgName', response.data.data.name)
-            // let userData = { currentWorkspace, ...response.data.data }
-          })
-      } catch (err) {
-        console.log(err)
-      }
-    } else {
-      console.log('YOU ARE NOT LOGGED IN, PLEASE LOG IN')
-    }
-  }, [])
+  //   if ((user && token) !== null) {
+  //     try {
+  //       axios
+  //         .get(`https://api.zuri.chat/organizations/${currentWorkspace}`, {
+  //           headers: {
+  //             Authorization: `Bearer ${token}`
+  //           }
+  //         })
+  //         .then(response => {
+  //           localStorage.setItem('orgName', response.data.data.name)
+  //           // let userData = { currentWorkspace, ...response.data.data }
+  //         })
+  //     } catch (err) {
+  //       console.log(err)
+  //     }
+  //   } else {
+  //     console.log('YOU ARE NOT LOGGED IN, PLEASE LOG IN')
+  //   }
+  // }, [])
 
-  const config = {
-    headers: {
-      Authorization: `Bearer ${token}`
-    }
+
+ const logout = () => {
+   window.location.href = '/signout'  
   }
-  const logout = () => {
-    axios({
-      method: 'post',
-      url: `https://api.zuri.chat/auth/logout`,
-      headers: {
-        Authorization: `Bearer ${token}`
-      }
-    })
-      .then(res => {
-        console.log(res)
-        window.location.href = '/signout'
-      })
-      .catch(err => {
-        console.error(err)
-      })
-  }
+
   const [pause, setPause] = useState(false)
   const [statusModal, setStatusModal] = useState(false)
 
@@ -157,19 +139,8 @@ const TopbarModal = ({ members }) => {
       {/* The section that shows the topbarprofile */}
       {showModal ? (
         <section className={styles.topbarModal}>
-          <div
-            id="overlay"
-            onClick={() => setShowModal(false)}
-            className={styles.membersModalOverlay}
-          />
           <div className={styles.sectionOne}>
-            <div
-              className={styles.oneLeft}
-              role="button"
-              onClick={() => {
-                userProfileImage !== '' && setProfilePicView(!profilePicView)
-              }}
-            >
+            <div className={styles.oneLeft}>
               <img
                 src={userProfileImage !== '' ? userProfileImage : defaultAvatar}
                 alt="profile-pic"
@@ -189,9 +160,8 @@ const TopbarModal = ({ members }) => {
           </div>
 
           <div className={styles.sectionTwo}>
-            {/* <StatusBadgeModal /> */}
-            <div className={styles.emoji}>{user?.status?.tag} </div>
-            <div className={styles.statusContent}>{user?.status?.text}</div>
+            <StatusBadgeModal />
+            <p className={styles.statusContent}>{user.status}</p>
           </div>
 
           <div className={styles.sectionThree}>
@@ -272,7 +242,6 @@ const TopbarModal = ({ members }) => {
           <div className={styles.sectionFive}>
             <p onClick={logout}>Sign out of Team Einstein workspace</p>
           </div>
-          {profilePicView && <ProfilePicView />}
         </section>
       ) : null}
     </>
